Update models to entity-based modules.
Update models, migrations and admin.
