--- traffic_control/fixtures/lifecycles.json
@@ -1,34 +1,0 @@
-[
-    {
-        "model": "traffic_control.lifecycle",
-        "pk": "6e768aff-0d92-4f65-b821-d76c4843a254",
-        "fields": {
-            "status": "IN_USE",
-            "description": "Pysyv\u00e4sti voimassa oleva"
-        }
-    },
-    {
-        "model": "traffic_control.lifecycle",
-        "pk": "9291623d-282a-405a-ad40-ea58f2d883f0",
-        "fields": {
-            "status": "TEMPORARILY_NOT_IN_USE",
-            "description": "Tilap\u00e4isesti pois k\u00e4yt\u00f6st\u00e4 oleva"
-        }
-    },
-    {
-        "model": "traffic_control.lifecycle",
-        "pk": "a1c25e6f-00b2-436f-a351-177ceae48641",
-        "fields": {
-            "status": "NOT_IN_USE",
-            "description": "Poistuva pysyv\u00e4 laite"
-        }
-    },
-    {
-        "model": "traffic_control.lifecycle",
-        "pk": "ad4c9c10-96af-43cc-9966-e898c744e515",
-        "fields": {
-            "status": "TEMPORARILY_IN_USE",
-            "description": "Tilap\u00e4isesti k\u00e4yt\u00f6ss\u00e4 oleva"
-        }
-    }
-]
